Cargo.toml handling is more flexible
In order to allow usage of helper libraries across all folders Cargo
allows specifying them via [lib]. This was removed on each build,
previously which no longer happens. These changes however mean that
Cargo is not as strictly controlled (now_lambda is not set to latest
always, however the [[bin]] is overwritten).

Cargo.toml is now also backed up and restored for builds, so that it
does not get updated for the users source control management.