Fix `now dev` and other changes (#19)
* Removed openssl installation

The openssl installation was breaking `now dev` on any environment
that isn't using `yum` as package manager. It seems sensible to remove
it and instead projects that require it can add it (along with package
management logic) into `build.sh`.

* Require main function in entrypoints

The replacement logic was breaking `now dev` as it was adding imports
and the main function each time `now dev` was run. To mitigate this, the
entrypoints now require the main function.

* Debugging matches official builders

Other builders are using @now/build-utils `debug` function to log debug
messages. It is hidden behind NOW_BUILD_DEBUG environment variable, so I
have also matched that behaviour for cargo verbosity. Without enabling,
cargo is now less verbose, however still shows compilation errors.

* Now only support version 3 of runtimes

* Added support for path segments

rustc does not support `[]` in crate names, hence cargo was failing to
build path segments. Replacing brackets with underscores for naming
fixes the issue.

* Check if rust exists before trying to install

This shaves off time as the installer no longer needs to download rustup
and try to install. Instead it checks before downloading.

* Cargo.toml handling is more flexible

In order to allow usage of helper libraries across all folders Cargo
allows specifying them via [lib]. This was removed on each build,
previously which no longer happens. These changes however mean that
Cargo is not as strictly controlled (now_lambda is not set to latest
always, however the [[bin]] is overwritten).

Cargo.toml is now also backed up and restored for builds, so that it
does not get updated for the users source control management.

* Updated tests for changes